reorganize otel code
